Added a separate Xcode project for testing OS X

+++ Tests/iOS Tests/SnapshotsTests.m
@@ -40,7 +40,7 @@ static NSUInteger numberOfIterations = 2500;
     dispatch_apply(n, lvldb_test_queue, ^(size_t i) {
         do {
             r = arc4random_uniform(5000);
-            key = [NSString stringWithFormat:@"%d", r];
+            key = [NSString stringWithFormat:@"%ld", (long)r];
         } while ([db objectExistsForKey:key]);
         
         value = @[@(r), @(i)];
